+++ packages/keyring-controller/jest.config.js
@@ -23,7 +23,4 @@ module.exports = merge(baseConfig, {
       statements: 99.18,
     },
   },
-
-  // These tests rely on the Crypto API
-  testEnvironment: '<rootDir>/jest.environment.js',
 });